simplify userfacing script trackingPipeline.py
The trained model will be in the same place all the time so it didn't
make sense for the model path to be an argument. Also the image
directory argument was changed to a folder name. It is assumed the user
will just place the folder for tracking in the main directory.