enable arbitrary input image dimensions
Prior to this, the cropping of the image prior to inference was
dictated by the user by entering dimensions. This is still a possible
argument but by default the image is cropped to the nearest dimensions
that are also divisible by 16. This is to make up for the pixels
lost during forward propagation of the network.